+++ node/messages/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2018"
 description = "Message types used by Subsystems"
+
+[dependencies]
+polkadot-primitives = { path = "../../primitives" }
